Add user authentication

+++ routes/task.js
@@ -3,19 +3,26 @@ var router = express.Router();
 
 const mysql = require('../mysql');
 const { paginateQuery } = require('../utils/sql');
+const { authenticationMiddleware } = require('../middleware');
 
 /* GET home page. */
+
+// app.use(authenticationMiddleware);
+
+router.use(authenticationMiddleware);
 
 router.get('/', (req, res, next) => {
   const page = req.query.page || 1;
   const limit = req.query.limit || 5;
+  const userId = req.user.id;
 
   try {
-    const queryCount = `SELECT COUNT(id) as count FROM todo`;
+    const queryCount = `SELECT COUNT(id) as count FROM todo WHERE user_id = ${userId}`;
     mysql(queryCount, count => {
       if(count){
         const totalCount = count[0].count;
-        const query = `SELECT id, title, due_date, done, user from todo`;
+        const query = `SELECT id, title, due_date, done, user from todo WHERE user_id = ${userId}`;
+
         mysql(paginateQuery(query, totalCount, page, limit), response => {
           const responseFormat = {
             count: totalCount,
@@ -24,8 +31,10 @@ router.get('/', (req, res, next) => {
             response: response
           }
             if(response.length >= 1){
+              res.status(200)
               res.json(responseFormat);
             }else{
+              res.status(404);
               res.send("Il n'y a rien dans la bd");
             }
           
@@ -40,12 +49,14 @@ router.get('/', (req, res, next) => {
 
 router.get('/:id', (req, res, next) => {
   const taskId = req.params.id;
+  const userId = req.user.id;
   try {
     const query = `SELECT * from todo WHERE id=${taskId}`;
     mysql(query, response => {
       if(response){
         const task = response.find((task) => task.id === Number(taskId));
-        if(task){
+        console.log(task)
+        if(task && task.user_id === userId){
           res.status(200);
           res.json(task);
         }else {
@@ -66,9 +77,10 @@ router.get('/:id', (req, res, next) => {
 router.post('/', (req, res, next) => {
 
   const done = req.body.done ? 1 : 0;
+  const userId = req.user.id;
 
   try {
-    const query = `Insert into todo (title, creation_date, due_date, done, description, user) Values ('${req.body.title}', '${req.body.creation_date}', '${req.body.due_date}', '${done}', '${req.body.description}', '${req.body.user}')`;
+    const query = `Insert into todo (title, creation_date, due_date, done, description, user_id) Values ('${req.body.title}', '${req.body.creation_date}', '${req.body.due_date}', '${done}', '${req.body.description}', '${userId}')`;
     mysql(query, response => {
       if(response){
         res.status(201)
@@ -86,18 +98,39 @@ router.post('/', (req, res, next) => {
 router.patch('/:id', (req, res, next) => {
   const taskId = req.params.id;
   const done = req.body.done ? 1 : 0;
+  const userId = req.user.id;
 
   try {
-    const query = `Update todo set title='${req.body.title}', creation_date='${req.body.creation_date}', due_date='${req.body.due_date}', done='${done}', description='${req.body.description}', user='${req.body.user}' where id=${taskId}`;
-    mysql(query, response => {
-      if(response){
-        res.status(200)
-        res.send("Tâche modifiée !")
+    const queryFind = `SELECT * from todo WHERE id=${taskId}`;
+    mysql(queryFind, responseFind => {
+      if(responseFind){
+        const task = responseFind.find((task) => task.id === Number(taskId));
+        console.log(task)
+        if(task && task.user_id === userId){
+          try {
+            const query = `Update todo set title='${req.body.title}', creation_date='${req.body.creation_date}', due_date='${req.body.due_date}', done='${done}', description='${req.body.description}', user='${req.body.user}' where id=${taskId}`;
+            mysql(query, response => {
+              if(response){
+                res.status(200)
+                res.send("Tâche modifiée !")
+              }else{
+                res.send("Une erreur est survenue.");
+              }
+            })
+          } catch(err) {
+            res.status(500)
+            res.send(err);
+          }
+        }else {
+          res.status(404);
+          res.send("L'identifiant ne correspond à aucun élément.");
+        }
       }else{
-        res.send("Une erreur est survenue.");
+        res.status(404);
+        res.send("L'identifiant ne correspond à aucun élément.");
       }
     })
-  } catch(err) {
+  }catch(err) {
     res.status(500)
     res.send(err);
   }
@@ -106,12 +139,13 @@ router.patch('/:id', (req, res, next) => {
 
 router.delete('/:id', (req, res, next) => {
   const taskId = req.params.id;
+  const userId = req.user.id;
   try {
     const query = `SELECT * from todo WHERE id=${taskId}`;
     mysql(query, response => {
       if(response){
         const task = response.find((task) => task.id === Number(taskId));
-        if(task){
+        if(task && task.user_id === userId){
           try {
             const query = `Delete from todo where id=${taskId}`;
             mysql(query, response => {
